pyproject.toml
[project]
name = "enum1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.2.1",
    "streamlit>=1.45.1",
]
